Fixing double-int problems in Player.h and Character.h

+++ Character.h
@@ -17,8 +17,8 @@ class Character
 {
 protected:
     const std::string characterName;	///< The character's name
-    int characterHP;	///< The character's HP
-    int characterDMG;	///< The character's DMG
+    double characterHP;	///< The character's HP
+    double characterDMG;	///< The character's DMG
     const double characterACD;	///< The character's ACD
 
 public:
